creates component for Vote
